requirements.txt
asgiref==3.7.2
Django==4.2.1
django-leaflet==0.29.0
et-xmlfile==1.1.0
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
